Completely finished implementation. Basic debuggings were done. Still need to execute more test cases. Also, test efficiency on large files

+++ include/encryption.h
@@ -5,7 +5,7 @@
 #include"node.h"
 
 typedef struct bitcode {
-  short numOfBits;
+  int numOfBits;
   unsigned char bits[256];
 } bitcode;
